Moved serialization to separate crate.
This avoids confusion over what serialization is.
