+++ README.md
@@ -60,21 +60,19 @@ with them from spark, thereby hindering robust integration tests.
 * `types`: physical type declaration (i.e. how things are represented in memory). So far unused.
 * `compression`: compression (e.g. Gzip)
 * `error`: errors declaration
-* `serialization`: convert from bytes to rust native typed buffers (`Vec<Option<T>>`).
 
-Note that `serialization` is not very robust. It serves as a playground 
-to understand the specification and how to serialize and deserialize pages.
+## Run integration tests
 
-## Run unit tests
-
-There are tests being run against parquet files generated by pyarrow. To ignore these,
-use `PARQUET2_IGNORE_PYARROW_TESTS= cargo test`. To run then, you will need to run 
+There are integration tests against parquet files generated by pyarrow.
+To run then, you will need to run
 
 ```bash
+cd integration-tests
 python3 -m venv venv
 venv/bin/pip install pip --upgrade
 venv/bin/pip install pyarrow==4
 venv/bin/python integration/write_pyarrow.py
+cargo test
 ```
 
 before. This is only needed once (per change in the `integration/write_pyarrow.py`).
@@ -101,18 +99,16 @@ let mut iter = get_page_iterator(&metadata, row_group, column, &mut file)?;
 /// A page. It is just (compressed) bytes at this point.
 let page = iter.next().unwrap().unwrap();
 println!("{:#?}", page);
-
-/// from here, we can do different things. One of them is to convert its buffers to native Rust.
-/// This consumes the page.
-use parquet2::serialization::native::page_to_array;
-let array = page_to_array(page, &descriptor).unwrap();
 ```
 
 ### How to implement page readers
 
 In general, the in-memory format used to consume parquet pages strongly influences how the pages should be deserialized. As such, this crate does not commit to a particular in-memory format. Consumers are responsible for converting pages to their target in-memory format.
 
-There is an implementation that uses the arrow format [here](https://github.com/jorgecarleitao/arrow2).
+This repository contains a serialization to a simple in-memory format in `array-tests`, that are
+used to validate integration with other implementations.
+
+There is also implementation for the arrow format [here](https://github.com/jorgecarleitao/arrow2).
 
 ### Higher Parallelism
