+++ integration-tests/.gitignore
@@ -1,2 +1,4 @@
 target
 Cargo.lock
+venv
+fixtures
