+++ integration-tests/README.md
@@ -1,0 +1,4 @@
+# Page reader and writer
+
+This crate contains an in-memory format used to test the parquet2
+crate in integration against pyarrow.
