--- src/serialization/mod.rs
@@ -1,2 +1,0 @@
-pub mod read;
-pub mod write;
